+++ .meta-updater/src/index.ts
@@ -106,6 +106,7 @@ async function updateTSConfig (
 ): Promise<object | null> {
   if (tsConfig == null) return tsConfig
   if (manifest.name === '@pnpm/tsconfig') return tsConfig
+  if (manifest.name === '@pnpm-private/typecheck') return tsConfig
   const relative = normalizePath(path.relative(context.workspaceDir, dir)) as ProjectId
   const importer = context.lockfile.importers[relative]
   if (!importer) return tsConfig
@@ -144,30 +145,13 @@ async function updateTSConfig (
     await writeJsonFile(path.join(dir, 'test/tsconfig.json'), {
       extends: '../tsconfig.json',
       compilerOptions: {
-        // The composite flag allows projects to be specified as references in
-        // other projects. Test projects should not be dependencies of other
-        // files and don't need composite set.
-        //
-        // Some tests perform a relative import into the "src" directory to test
-        // non-publicly exported idenfiers.
-        //
-        //   import { foo } from "../src/foo"
-        //
-        // Instead of:
-        //
-        //   import { foo } from "@pnpm/example"
-        //
-        // The relative "../src" imports would error if composite is enabled
-        // since composite would require files in "src" to be part of the test
-        // project.
-        composite: false,
-        noEmit: true,
-
+        noEmit: false,
+        outDir: '../test.lib',
         rootDir: '.'
       },
       include: [
         '**/*.ts',
-        path.relative(testDir, path.join(context.workspaceDir, '__typings__/**/*.d.ts')),
+        normalizePath(path.relative(testDir, path.join(context.workspaceDir, '__typings__/**/*.d.ts'))),
       ],
       references: (tsConfig as any)?.compilerOptions?.composite === false
         // If composite is explicitly set to false, we can't add the main
@@ -184,7 +168,7 @@ async function updateTSConfig (
         // methods to be defensive against future changes to testDir, dir, or
         // relPath.
         ? linkValues.map(relPath => ({
-          path: path.relative(testDir, path.join(dir, relPath))
+          path: normalizePath(path.relative(testDir, path.join(dir, relPath)))
         }))
 
         // If the main project is composite (the more common case), we can
@@ -205,7 +189,7 @@ async function updateTSConfig (
       include: [
         'src/**/*.ts',
         'test/**/*.ts',
-        path.relative(dir, path.join(context.workspaceDir, '__typings__/**/*.d.ts')),
+        normalizePath(path.relative(dir, path.join(context.workspaceDir, '__typings__/**/*.d.ts'))),
       ],
     }, { indent: 2 })
   ])