+++ cli/default-reporter/src/reporterForClient/reportSkippedOptionalDependencies.ts
@@ -12,7 +12,8 @@ export function reportSkippedOptionalDependencies (
     filter((log) => Boolean(log['prefix'] === opts.cwd && log.parents && log.parents.length === 0)),
     map((log) => Rx.of({
       msg: `info: ${
-        log.package['id'] || log.package.name && (`${log.package.name}@${log.package.version}`) || log.package['pref']
+        // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
+        log.package.id || log.package.name && (`${log.package.name}@${log.package.version}`) || log.package.pref
       } is an optional dependency and failed compatibility check. Excluding it from installation.`,
     }))
   )