+++ config/normalize-registries/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/normalize-registries
 
+## 6.0.5
+
+### Patch Changes
+
+- Updated dependencies [cb006df]
+  - @pnpm/types@12.0.0
+
 ## 6.0.4
 
 ### Patch Changes