+++ lockfile/walker/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pnpm/lockfile-walker
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [cb006df]
+  - @pnpm/lockfile.types@1.0.1
+  - @pnpm/types@12.0.0
+  - @pnpm/dependency-path@5.1.4
+
 ## 1.0.0
 
 ### Major Changes