+++ network/fetch/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/fetch
 
+## 8.0.5
+
+### Patch Changes
+
+- Updated dependencies [cb006df]
+  - @pnpm/types@12.0.0
+  - @pnpm/core-loggers@10.0.5
+
 ## 8.0.4
 
 ### Patch Changes