+++ reviewing/plugin-commands-licenses/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @pnpm/plugin-commands-licenses
 
+## 4.1.15
+
+### Patch Changes
+
+- @pnpm/cli-utils@4.0.1
+
+## 4.1.14
+
+### Patch Changes
+
+- Updated dependencies [26b065c]
+- Updated dependencies [26b065c]
+  - @pnpm/cli-utils@4.0.0
+  - @pnpm/config@21.8.0
+
+## 4.1.13
+
+### Patch Changes
+
+- Updated dependencies [98c8bd6]
+- Updated dependencies [d20eed3]
+  - @pnpm/cli-utils@3.1.7
+  - @pnpm/config@21.7.0
+  - @pnpm/lockfile.fs@1.0.1
+  - @pnpm/license-scanner@3.1.11
+
 ## 4.1.12
 
 ### Patch Changes